lots of updates, one being users and the ability to login
Changing the users table to use blowfish for the password. Also adding a remember_token column to
the users table so the users can login permenantly. Added the login and logout routes along with
views to allow the user to login or logout of the site.

Added some sections to the layout so things can be added to the logo or navigation areas easily
from other views.