moving the blog seeder to be separate and updating the entry views
The blog seeder should be a separate seeder for wordpress dumps. I should also find a way to parse a file
based on an argument passed in. This way the portfolio can have multiple seeders for the blog from different
blog software.

Changed the layout of the entry so images that are placeholders will show a bit more formatted. Also makes
it easier for templates to theme things later down the road.